requirements/requirements.txt
django-environ
Django
djangorestframework
psycopg2
django-filter